// block-performance-1.0.0/src/library/cloudman/ssh-apt-install/type.ts
export interface Input {
    username: string;
    address: string;
    ssh_auth_key: { os?: boolean; store?: boolean; file?: string };
    noninteractive?: boolean;
    app_list: string[];
}

export interface OutputOk {}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}